+++ src/__tests__/button_filterByLabel/labels_noneSelected.test.tsx
@@ -9,7 +9,7 @@ jest.mock('next/navigation', () => ({
 import Labels from '@/app/labels';
 import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
-import { labels } from '../testcases/labels';
+import { labels } from '../__testcases__/labels';
 
 describe("Labels when no label is selected", () => {
   Object.entries(labels.selected.none).map(([buttonName, { href }]: [string, { href: string }], index: number) => {
